range-diff: clarify --creation-factor=<factor>
The value is not a per-cent that ranges from 0 to 100.  The SYNOPSIS
section gets it right, but the body of the documentation said "percent"
which confused readers.

While we are at it, rephrase "smaller one" that corresponds to
"larger value" earlier in the sentence to "smaller value" to be more
explicit, to avoid misleading eyes of the readers to an unrelated "a
large change" nearby.

Signed-off-by: Junio C Hamano <gitster@pobox.com>